+++ CHANGELOG.md
@@ -20,6 +20,10 @@
   [Jason Vasquez](https://github.com/jasonvasquez)
   [Xcodeproj#320](https://github.com/CocoaPods/Xcodeproj/pull/320)
 
+* Added basic Xcodeproj::Workspace APIs to support groups
+  [David Parton](https://github.com/dparton)
+  [Xcodeproj#322](https://github.com/CocoaPods/Xcodeproj/pull/322)
+
 ##### Bug Fixes
 
 * Avoid duplicating settings with multiple values in common when merging.